lib: collisions: rm lib and move code to corresponding obstacles classes
