--- lib/collisions/Makefile.dep
@@ -1,1 +1,0 @@
-USEMODULE += obstacles
