--- lib/collisions/Makefile.include
@@ -1,2 +1,0 @@
-USEMODULE_INCLUDES_collisions := $(LAST_MAKEFILEDIR)/include
-USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_collisions)
