+++ lib/obstacles/Makefile.dep
@@ -1,4 +1,3 @@
 USEMODULE += cogip_defs
-USEMODULE += collisions
 USEMODULE += trigonometry
 USEMODULE += tracefd
